backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
celery
django-celery-beat